pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wdgrid"
keywords = ["nicegui", "wikidata","sync","ui"]
description = "wikidata grid display and sync"
authors = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"}
]
maintainers = [
  {name = "Wolfgang Fahl", email = "wf@bitplan.com"},
]
readme = "README.md"
license= "Apache-2.0"
dependencies = [
	# https://github.com/WolfgangFahl/nicegui_widgets
	"ngwidgets>=0.14.0",
	# https://pypi.org/project/pyLodStorage/
	"pyLodStorage>=0.12.2",
	#https://pypi.org/project/tabulate/
	"tabulate>=0.9.0",
	# https://pypi.org/project/py-ez-wikidata/
	"py-ez-wikidata>=0.2.0"
]

requires-python = ">=3.9"
classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: Apache Software License"
]
dynamic = ["version"]

[tool.hatch.version]
path = "wd/__init__.py"

[project.urls]
Home = "https://github.com/WolfgangFahl/wdgrid"
Documentation = "https://wiki.bitplan.com/index.php/Wdgrid"
Source = "https://github.com/WolfgangFahl/wdgrid"

[project.optional-dependencies]
test = [
  "green",
]

[tool.hatch.build.targets.wheel]
only-include = ["wd"]

[tool.hatch.build.targets.wheel.sources]
"wd" = "wd"

[project.scripts]
wdgrid = "wd.wdgrid_cmd:main"
